+++ tst/unit/services/product/TestGetAllProducts.java
@@ -1,0 +1,15 @@
+package unit.services.product;
+
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+import unit.TestVioletServices;
+
+public class TestGetAllProducts extends TestVioletServices {
+
+    @Test
+    public void getAllProducts() {
+        Assert.assertEquals(productServices.all().getBody().size(), productRepository.count());
+    }
+
+}
